DAO-181 Move menu icons into ~/components/icons folder
