+++ src/components/icons/dashboard-icon.tsx
