+++ src/components/icons/proposals-icon.tsx
