+++ src/components/menu/menu.tsx
@@ -3,9 +3,9 @@ import { NavLink, useLocation } from 'react-router-dom';
 import { useChainData } from '../../chain-data';
 import { images } from '../../utils';
 import classNames from 'classnames';
-import DashboardIcon from './dashboard-icon';
-import ProposalsIcon from './proposals-icon';
-import HistoryIcon from './history-icon';
+import DashboardIcon from '../icons/dashboard-icon';
+import ProposalsIcon from '../icons/proposals-icon';
+import HistoryIcon from '../icons/history-icon';
 import PoliciesIcon from '../icons/policies-icon';
 import ClaimsIcon from '../icons/claims-icons';
 import TrackerIcon from '../icons/tracker-icon';
